158 - implemented PDF generation on the backend and moved relevant functionality to the order confirmation page on the frontend
