+++ frontend/src/app/app-routing.module.ts
@@ -6,6 +6,7 @@ import { ApplicationSubmittedComponent } from './application-forms/application-s
 import { AccessControlService } from './_services/access-control.service';
 import { AdminAccessControlService } from './_services/admin-access-control.service';
 import { ChristmasTreePermitResolver } from './application-forms/tree-application-form/christmas-tree-permit-resolver.service';
+import { FirewoodPermitResolver } from './application-forms/tree-application-form/firewood-permit-resolver.service';
 import { DummyComponent } from './print-permit-dummy-page/dummy.component';
 import { ForestResolver } from './trees/forests/tree-guidelines/forest-resolver.service';
 import { ForestsResolver } from './trees/forests/forest-finder/forests-resolver.service';
@@ -371,6 +372,9 @@ const appRoutes: Routes = [
             children: [
               {
                 path: '',
+                resolve: {
+                  permit: FirewoodPermitResolver
+                },
                 component: OrderConfirmationComponent
               }
             ]
