+++ frontend/src/app/firewood/forests/firewood-guidelines/firewood-guidelines.component.html
@@ -21,9 +21,6 @@
       <p>You can buy firewood permits online or in-person at one of <a href="#prohibited">our district offices</a>.</p>
       <button #staticBuyButton [routerLink]="['/firewood/forests/', forest.forestAbbr, 'buy']" class="usa-button margin-top-1" id="static-buy-permit-link" tabIndex="-1">Buy permit</button>
 
-      <span (click)="printPDF()">Print PDF</span>
-      <span (click)="emailPDF()">Email PDF</span>
-
       <!-- where to find firewood -->
       <app-where-to-find></app-where-to-find>
