+++ frontend/src/app/firewood/forests/order-confirmation/order-confirmation.component.html
@@ -8,7 +8,7 @@
 
     <div class="confirmation-details-container margin-top-2">
       <p>Your firewood permit has been purchased.</p>
-      <p>Your permit, load tags and required documents have been sent to <strong>email@address.com</strong>. If for some reason this email address is incorrect, please print your documents now.</p>
+      <p>Your permit, load tags and required documents have been sent to <strong>{{permit.emailAddress}}</strong>. If for some reason this email address is incorrect, please print your documents now.</p>
       <p><strong>You will need to have your permit, load tags and required documents with you when collecting firewood for your permit to be valid.</strong> Your load tags will have instructions for how to fill it out and attach it to your load.</p>
     </div>
 
@@ -19,18 +19,16 @@
         <li>Your load tags</li>
         <li>Cutting map</li>
       </ul>
-      <input id="includeRules" type="checkbox" name="includeRules" [checked]="includeRules" (change)="includeRules = !includeRules">
-      <label for="includeRules">Include a copy of the firewood cutting regulations</label>
-      <button id="tree-print-permit" class="usa-button-primary usa-button-big" type="button" [disabled]="isPermitExpired" (click)="printPermit()">
+      <button id="tree-print-permit" class="usa-button-primary usa-button-big" type="button" [disabled]="isPermitExpired" (click)="printPDF()">
         Print now
       </button>
-      <p>These documents have also been emailed to <strong>email@address.com</strong></p>
+      <p>These documents have also been emailed to <strong>{{permit.emailAddress}}</strong></p>
     </div>
 
     <div *ngIf="forest" id='tree-order-summary-container' class="confirmation-details-container clear margin-top-2">
       <div *ngIf="isPermitExpired" class="usa-alert usa-alert-error permit-expiration-warning">
         <div class="usa-alert-body">
-          <!-- <h3 class="usa-alert-heading">This permit was issued on {{permit.transactionDate | date: 'shortDate'}} and expired on {{permit.expirationDate | date: 'shortDate'}}</h3> -->
+          <h3 class="usa-alert-heading">This permit was issued on {{permit.transactionDate | date: 'shortDate'}} and expired on {{permit.expirationDate | date: 'shortDate'}}</h3>
         </div>
       </div>
       <h2 class="usa-heading">Order summary</h2>
@@ -44,8 +42,7 @@
       <div class="firewood-order-detail usa-width-one-third">
         <dl>
           <dt><strong>Quantity</strong></dt>
-          <dd>4</dd>
-          <!-- <dd>{{permit.quantity}}</dd> -->
+          <dd>{{permit.quantity}}</dd>
         </dl>
       </div>
     </div>
@@ -54,20 +51,23 @@
       <h2 class="usa-heading">Order details</h2>
       <hr class="header-divider">
       <table class="tree-confirmation-table">
+        <div class="firewood-order-detail usa-width-one-third">
+          <dl>
+            <dt><strong>Quantity</strong></dt>
+            <dd>{{permit.quantity}}</dd>
+          </dl>
+        </div>
         <tr class="firewood-order-detail">
           <td><strong>Permit number:</strong></td>
-          <td>234</td>
-          <!-- <td>{{permit.permitNumber | uppercase}}</td> -->
+          <td>{{permit.permitNumber | uppercase}}</td>
         </tr>
         <tr class="firewood-order-detail">
           <td><strong>Name:</strong></td>
-          <td>Firstname Lastname</td>
-          <!-- <td>{{permit.firstName}} {{permit.lastName}}</td> -->
+          <td>{{permit.firstName}} {{permit.lastName}}</td>
         </tr>
         <tr class="firewood-order-detail">
           <td><strong>Payment:</strong></td>
-          <td>$20</td>
-          <!-- <td>${{permit.totalCost}}</td> -->
+          <td>${{permit.totalCost}}</td>
         </tr>
         <tr class="firewood-order-detail">
           <td><strong>Transaction method:</strong></td>
@@ -75,8 +75,7 @@
         </tr>
         <tr class="firewood-order-detail">
           <td><strong>Transaction date:</strong></td>
-          <td>1/2/30</td>
-          <!-- <td>{{permit.transactionDate | date: 'shortDate'}}</td> -->
+          <td>{{permit.transactionDate | date: 'shortDate'}}</td>
         </tr>
       </table>
     </div>
@@ -84,10 +83,8 @@
     <div id="tree-additional-info-container" class="confirmation-details-container usa-width-one-whole margin-top-2 margin-bottom-2">
       <h2 class="usa-heading">Additional information</h2>
       <hr class="header-divider">
-      <p>Your permit, load tag and required documents were emailed to you at <strong>email@address.com.</strong></p>
-      <!-- <h3>Your permit and the Christmas tree cutting rules and guidelines were emailed to you at {{permit.emailAddress}}.</h3> -->
+      <p>Your permit, load tag and required documents were emailed to you at <strong>{{permit.emailAddress}}</strong></p>
       <p>Return to the <a [routerLink]="'/firewood/forests/' + forest.forestAbbr">{{forest.forestName}} Firewood permit website.</a></p>
-      <!-- <p *ngIf="forest">Return to the <a [routerLink]="'/christmas-trees/forests/' + forest.forestAbbr">{{forest.forestName}} Christmas tree permit website.</a></p> -->
     </div>
   </div>
