+++ server/src/email/email-util.es6
@@ -1,5 +1,3 @@
-
-
 /**
  * Module for common controllers
  * @module email/email-util
@@ -11,6 +9,7 @@ const logger = require('../services/logger.es6');
 const vcapConstants = require('../vcap-constants.es6');
 const emailTemplates = require('./email-templates.es6');
 const htmlTemplate = require('./assets/html-template.es6');
+const fs = require('fs');
 
 
 const emailUtil = {};
@@ -115,7 +114,12 @@ emailUtil.sendPermit = async (data) => {
   };
 
   let emailResponse = await emailUtil.transporter.sendMail(mailOptions)
-  console.dir(emailResponse)
+  try {
+    fs.unlinkSync('permit.pdf');
+    console.log('successfully deleted permit.pdf');
+  } catch (err) {
+    console.log('error deleting permit.pdf')
+  }
   return emailResponse
 
 }
